Update SN2011fe.json
Pardon the delay.

OK so what I've done here is add in the display attribute assuming the spectrum is shown between 3320 and 9910 Angstroms for the first spectrum in need of trimming. This should be used to cut/grey-out everything blueward of 3320 Ang and redward of 9910 Ang. Alternatively I could create two display windows that specify a window to cut, one in the blue, one in the red. 

Side notes: 
--I've specified the filename for each spectrum in order to identify which display window goes where. Thoughts?
--In order to get the spectrum to show up with a better scaling, is it possible to normalize wrt the "new" spectrum's wavelength range?
--We could instead do "blue", "red", "cosmic/spike" to cut any cosmic ray spikes and/or else.