SN2002au's claimedtype from Transient Name Server,2002IAUC.7823....3J,Latest Supernovae,2012A&A...538A.120L marked as being erroneous.
So this is an interesting one. For SN 2002au, the type is claimed to be Ia and a IIb. The spectrum of 2002au shows the IIb classification to be more accurate, but when I go to the IAU circular that classified it as a Ia, instead I find that the IAUC does not classify 2002au [at all], but instead classified a different event (2002as) as a II.