SN2003aa's claimedtype from 1989A&AS...81..421B,Transient Name Server,2003IAUC.8063....3C,Latest Supernovae,2012A&A...538A.120L marked as being erroneous.
I'm going to change "claimedtype" to reflect IAUC 8064 instead (a Ic, not a Ia).

SUPERNOVA 2003aa IN NGC 3367
     Further to IAUC 8063, T. Matheson, P. Challis, and R. Kirshner
report that a spectrum (range 370-750 nm) of SN 2003aa (cf. IAUC
8063), obtained by P. Berlind with the 1.5-m reflector on Feb. 2.31
UT, shows it to be a type-Ic supernova.  The spectrum is similar to
that of SN 1994I, four days before maximum (Filippenko et al. 1995,
Ap.J. 450, L11).