Update SN2005fk.json
SN2005fk received as Ic-BL classification in Taddia et al. (2015). I have added this classification as well as the reference. When I clicked "edit" on the homepage I was taken to creating this new .json page, though, instead of editing an existing one -- I am not sure why that is.